requirements.txt
streamlit==1.50.0
tensorflow-cpu==2.20.0
pillow==10.4.0
numpy==2.3.4
